pyproject.toml
[project]
name = "agentforge"
version = "0.1.0"
description = "An opinionated, open-source, code-first Python stack for building AI agents."
requires-python = ">=3.12"
dependencies = [
    # Agent framework
    "pydantic-ai[groq]>=0.1.0",
    # API layer
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Scheduling
    "apscheduler>=4.0.0",
    # Database
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Data validation
    "pydantic>=2.10.0",
    # HTTP client (collectors only)
    "httpx>=0.28.0",
    # Observability
    "langfuse>=2.0.0,<3.0.0",
    # Environment
    "python-dotenv>=1.0.0",
    # YouTube collector dependencies
    "google-api-python-client>=2.150.0",
    "youtube-transcript-api>=0.6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ---------------------------------------------------------------------------
# Ruff — replaces flake8, black, isort
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line too long — handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
